+++ src/ast/irbuilderbpf.cpp
@@ -662,21 +662,22 @@ Value *IRBuilderBPF::CreateWriteMapValueAllocation(const SizedType &value_type,
 
 Value *IRBuilderBPF::CreateVariableAllocationInit(const SizedType &value_type,
                                                   const std::string &name,
-                                                  const Location &loc)
+                                                  const Location &loc,
+                                                  bool force_scratch)
 {
   // Hoist variable declaration and initialization to entry point of
   // probe/subprogram. While we technically do not need this as variables
   // are properly scoped, it eases debugging and is consistent with previous
   // stack-only variable implementation.
   Value *alloc;
-  hoist([this, &value_type, &name, &loc, &alloc] {
-    alloc = createAllocation(bpftrace::globalvars::VARIABLE_BUFFER,
-                             GetType(value_type),
-                             name,
-                             loc,
-                             [](AsyncIds &async_ids) {
-                               return async_ids.variable();
-                             });
+  hoist([this, &value_type, &name, &loc, &alloc, force_scratch] {
+    alloc = createAllocation(
+        bpftrace::globalvars::VARIABLE_BUFFER,
+        GetType(value_type),
+        name,
+        loc,
+        [](AsyncIds &async_ids) { return async_ids.variable(); },
+        force_scratch);
     CreateAllocationInit(value_type, alloc);
   });
   return alloc;
@@ -700,11 +701,12 @@ Value *IRBuilderBPF::createAllocation(
     llvm::Type *obj_type,
     const std::string &name,
     const Location &loc,
-    std::optional<std::function<size_t(AsyncIds &)>> gen_async_id_cb)
+    std::optional<std::function<size_t(AsyncIds &)>> gen_async_id_cb,
+    bool force_scratch)
 {
   const auto obj_size = module_.getDataLayout().getTypeAllocSize(obj_type);
   const auto on_stack_limit = bpftrace_.config_->on_stack_limit;
-  if (obj_size > on_stack_limit) {
+  if (force_scratch || obj_size > on_stack_limit) {
     return createScratchBuffer(global_var_name,
                                loc,
                                gen_async_id_cb ? (*gen_async_id_cb)(async_ids_)