+++ src/ast/irbuilderbpf.h
@@ -149,7 +149,8 @@ public:
                                        const Location &loc);
   Value *CreateVariableAllocationInit(const SizedType &value_type,
                                       const std::string &name,
-                                      const Location &loc);
+                                      const Location &loc,
+                                      bool force_scratch = false);
   Value *CreateMapKeyAllocation(const SizedType &value_type,
                                 const std::string &name,
                                 const Location &loc);
@@ -300,7 +301,8 @@ private:
                           const std::string &name,
                           const Location &loc,
                           std::optional<std::function<size_t(AsyncIds &)>>
-                              gen_async_id_cb = std::nullopt);
+                              gen_async_id_cb = std::nullopt,
+                          bool force_scratch = false);
   void CreateAllocationInit(const SizedType &stype, Value *alloc);
   Value *createScratchBuffer(std::string_view global_var_name,
                              const Location &loc,